# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ingesting, storing, and querying log data with Grok pattern support",
        "title": "LogSonic API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/grok": {
            "get": {
                "description": "Create, read, update, and delete Grok patterns for log parsing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grok"
                ],
                "summary": "Manage Grok patterns",
                "parameters": [
                    {
                        "description": "Grok pattern definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GrokPatternRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response with patterns",
                        "schema": {
                            "$ref": "#/definitions/types.GrokPatternResponse"
                        }
                    },
                    "201": {
                        "description": "Pattern created successfully",
                        "schema": {
                            "$ref": "#/definitions/types.GrokPatternResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format or missing required fields",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pattern not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create, read, update, and delete Grok patterns for log parsing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grok"
                ],
                "summary": "Manage Grok patterns",
                "parameters": [
                    {
                        "description": "Grok pattern definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GrokPatternRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response with patterns",
                        "schema": {
                            "$ref": "#/definitions/types.GrokPatternResponse"
                        }
                    },
                    "201": {
                        "description": "Pattern created successfully",
                        "schema": {
                            "$ref": "#/definitions/types.GrokPatternResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format or missing required fields",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pattern not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Create, read, update, and delete Grok patterns for log parsing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grok"
                ],
                "summary": "Manage Grok patterns",
                "parameters": [
                    {
                        "description": "Grok pattern definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GrokPatternRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response with patterns",
                        "schema": {
                            "$ref": "#/definitions/types.GrokPatternResponse"
                        }
                    },
                    "201": {
                        "description": "Pattern created successfully",
                        "schema": {
                            "$ref": "#/definitions/types.GrokPatternResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format or missing required fields",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pattern not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/info": {
            "get": {
                "description": "Retrieve detailed information about the system, storage, and application",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get system and storage information",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Force a cache refresh if set to true",
                        "name": "refresh",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SystemInfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ingest/end": {
            "post": {
                "description": "End the specified log ingest session and cleanup its resources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "End log ingest session",
                "parameters": [
                    {
                        "description": "Session end request with session_id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IngestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.IngestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ingest/file": {
            "post": {
                "description": "Start ingesting a log file the server can access (absolute path; gzip and zstd are detected by magic bytes) into an existing ingest session, optionally with its rotated siblings. Accepted immediately (202); the read runs in the background. Poll GET /ingest/jobs or listen for \"ingest_progress\" on GET /live/events for progress and the final state. Call POST /ingest/end only after the job reaches a terminal state (done|cancelled|error) -- ending the session while the job is still running fails its next batch with an invalid-session error.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "Ingest a file by path",
                "parameters": [
                    {
                        "description": "Session and absolute path",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.IngestFileRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/types.IngestFileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.GrokPatternRequest": {
            "description": "Request structure for creating or updating Grok patterns",
            "type": "object",
            "properties": {
                "custom_patterns": {
                    "description": "Map of custom pattern definitions used by this pattern",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "description": {
                    "description": "Human-readable description of the pattern",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the Grok pattern",
                    "type": "string"
                },
                "pattern": {
                    "description": "The Grok pattern string",
                    "type": "string"
                },
                "priority": {
                    "description": "Priority of the pattern (higher numbers are matched first)",
                    "type": "integer"
                },
                "timestamp_config": {
                    "description": "TimestampConfig is the user's preferred resolution for this\npattern. Persisted in a logsonic-side file (log2grok's library\nschema doesn't carry it) and re-applied on subsequent imports.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/timeresolve.Resolution"
                        }
                    ]
                }
            }
        },
...
        "types.IngestRequest": {
            "type": "object",
            "properties": {
                "logs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
...
    }
}